Example improvements
Moved library installs and imports to separate cells at the top of the notebook. To do: look through the notebook to identify any more pip installs or library imports and move them to their respective cell.

Specified the output path in the prd_dash_lab Unity Catalog.

Added the "output_path" variable to the f-string which saves the *_species_distribution.png output files. To do: follow the same pattern to add the output_path variable in all other necessary lines.